optimize seo, accessibility, and best practices to green
