+++ next.config.mjs
@@ -4,11 +4,11 @@ const nextConfig = {
   
   async redirects() {
     return [
-      {
-        source: '/landingpage/landing-page',
-        destination: '/',
-        permanent: true,
-      },
+      // {
+      //   source: '/landingpage/landing-page',
+      //   destination: '/',
+      //   permanent: true,
+      // },
       {
         source: '/ingredients',
         destination: '/ingredients/list',
