+++ package.json
@@ -2,6 +2,7 @@
   "name": "what-can-i-cook",
   "version": "0.1.0",
   "private": true,
+  "sideEffects": false,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
