+++ src/pages/_app.js
@@ -1,5 +1,16 @@
 import "@/styles/globals.css";
+import Head from "next/head";
 
-export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+const App = ({ Component, pageProps }) => {
+  return (
+    <>
+      <Head>
+        <title>What Can I Cook?</title>
+        <meta name="description" content="Free Web tutorials"></meta>
+      </Head>
+      <Component {...pageProps} />;
+    </>
+  )
 }
+
+export default App;
