+++ src/pages/_document.js
@@ -1,6 +1,6 @@
 import { Html, Head, Main, NextScript } from "next/document";
 
-export default function Document() {
+const Document = () => {
   return (
     <Html lang="en">
       <Head>
@@ -26,3 +26,5 @@ export default function Document() {
     </Html>
   );
 }
+
+export default Document;
