+++ tailwind.config.js
@@ -67,6 +67,7 @@ module.exports = {
         		xl: '1280px',
         		'2xl': '1536px',
       		},
+			purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   		}
   	},
   	plugins: [require("tailwindcss-animate")],
